+++ plugins/repository-azure/src/main/java/org/opensearch/repositories/azure/AzureStorageService.java
@@ -32,6 +32,7 @@
 
 package org.opensearch.repositories.azure;
 
+import com.azure.identity.ManagedIdentityCredentialBuilder;
 import io.netty.channel.EventLoopGroup;
 import io.netty.channel.nio.NioEventLoopGroup;
 import io.netty.util.concurrent.Future;
@@ -48,6 +49,7 @@ import com.azure.core.http.policy.HttpPipelinePolicy;
 import com.azure.core.util.Configuration;
 import com.azure.core.util.Context;
 import com.azure.core.util.logging.ClientLogger;
+import com.azure.identity.ManagedIdentityCredentialBuilder;
 import com.azure.storage.blob.BlobServiceClient;
 import com.azure.storage.blob.BlobServiceClientBuilder;
 import com.azure.storage.blob.models.ParallelTransferOptions;
@@ -64,9 +66,11 @@ import org.opensearch.common.settings.SettingsException;
 import org.opensearch.common.unit.ByteSizeUnit;
 import org.opensearch.common.unit.ByteSizeValue;
 import org.opensearch.common.unit.TimeValue;
+import org.opensearch.common.Strings;
 
 import java.net.Authenticator;
 import java.net.PasswordAuthentication;
+import java.net.URI;
 import java.net.URISyntaxException;
 import java.security.InvalidKeyException;
 import java.time.Duration;
@@ -160,10 +164,27 @@ public class AzureStorageService implements AutoCloseable {
         return new Tuple<>(state.getClient(), () -> buildOperationContext(azureStorageSettings));
     }
 
+    public StorageEndpoint getStorageBlobEndpoint(final AzureStorageSettings settings) {
+        try {
+            String endpointSuffix = settings.getEndpointSuffix();
+            if (!Strings.hasText(endpointSuffix)) {
+                endpointSuffix = "core.windows.net";
+            }
+            final URI primaryBlobEndpoint = new URI("https://" + settings.getAccount() + ".blob." + endpointSuffix);
+            final URI secondaryBlobEndpoint = new URI("https://" + settings.getAccount() + "-secondary.blob." + endpointSuffix);
+            return new StorageEndpoint(primaryBlobEndpoint, secondaryBlobEndpoint);
+        } catch (URISyntaxException var14) {
+            throw logger.logExceptionAsError(new RuntimeException(var14));
+        }
+    }
+
     private ClientState buildClient(AzureStorageSettings azureStorageSettings, BiConsumer<HttpRequest, HttpResponse> statsCollector)
         throws InvalidKeyException, URISyntaxException {
         final BlobServiceClientBuilder builder = createClientBuilder(azureStorageSettings);
-
+        // When managed identity is enabled, no connection string will be generated, need to declare the primary uri
+        if (azureStorageSettings.usesManagedIdentityCredential()) {
+            builder.endpoint(getStorageBlobEndpoint(azureStorageSettings).getPrimaryUri());
+        }
         final NioEventLoopGroup eventLoopGroup = new NioEventLoopGroup(new NioThreadFactory());
         final NettyAsyncHttpClientBuilder clientBuilder = new NettyAsyncHttpClientBuilder().eventLoopGroup(eventLoopGroup);
 
@@ -217,8 +238,13 @@ public class AzureStorageService implements AutoCloseable {
      * https://github.com/Azure/azure-sdk-for-java/blob/main/sdk/storage/azure-storage-blob/migrationGuides/V8_V12.md#miscellaneous
      */
     private BlobServiceClientBuilder applyLocationMode(final BlobServiceClientBuilder builder, final AzureStorageSettings settings) {
-        final StorageConnectionString storageConnectionString = StorageConnectionString.create(settings.getConnectString(), logger);
-        final StorageEndpoint endpoint = storageConnectionString.getBlobEndpoint();
+        final StorageEndpoint endpoint;
+        if (settings.usesManagedIdentityCredential()) {
+            endpoint = getStorageBlobEndpoint(settings);
+        } else {
+            final StorageConnectionString storageConnectionString = StorageConnectionString.create(settings.getConnectString(), logger);
+            endpoint = storageConnectionString.getBlobEndpoint();
+        }
 
         if (endpoint == null || endpoint.getPrimaryUri() == null) {
             throw new IllegalArgumentException("connectionString missing required settings to derive blob service primary endpoint.");
@@ -250,6 +276,9 @@ public class AzureStorageService implements AutoCloseable {
 
     private static BlobServiceClientBuilder createClientBuilder(AzureStorageSettings settings) throws InvalidKeyException,
         URISyntaxException {
+        if (settings.usesManagedIdentityCredential()) {
+            return new BlobServiceClientBuilder().credential(new ManagedIdentityCredentialBuilder().build());
+        }
         return new BlobServiceClientBuilder().connectionString(settings.getConnectString());
     }